AIRFLOW_DOCKER/requirements.txt
pyspark==3.3.2
delta-spark==2.3.0
pandas==2.2.3
pyarrow==19.0.1
google-cloud-storage=3.1.0
google-cloud-bigquery==3.31.0